ci: repair PR #2 CI - deps, clippy, wire-format, renode epoch, clang-tidy
Getting the upstream PR green across every red check:

- Python: pytest-cov was invoked by CI but never declared; added to
  dev extras + uv.lock relocked.
- Rust: clippy useless_borrows_in_formatting in lichen-cli (writeln!
  with redundant &*).
- Wire Format Roundtrip: the sample generator imported nonexistent
  APIs (lichen.schc.rules.compress, lichen.coap.message) and never
  ran; rewritten against the real LichenFrame and aiocoap APIs.
- Renode Boot+TX: SOURCE_DATE_EPOCH was a literal '0', which the
  build-epoch policy rejects (range 1..2^31-1); use the commit
  timestamp like ci.yml.
- clang-tidy: cleared all 23 findings the enforcing gate reported.
  Real fixes: guarded a NULL-source memcpy for empty OSCORE kid (UB);
  propagated a _Nonnull contract in coap_oscore; a defensive NULL
  guard before Schnorr verify; pointed the RX ctx's _Nonnull peer_eui64
  at a valid zero buffer instead of NULL; removed dead stores in
  router/gradient/link_tx; removed redundant forward decls; explicit
  casts and (void) on discarded snprintf/return values. Curated out
  clang-analyzer-optin.core.EnumCastOutOfRange with rationale (fires on
  the idiomatic net_l2_flags OR that Zephyr's own L2 drivers use).

Verified locally: Python 2636 passed + ruff clean; Rust clippy (CI's
exact flags) + fmt clean; twister native_sim 47/47; clang-tidy clean on
all changed files.

+++ lichen/subsys/lichen/app_interface/app_interface.c
@@ -94,7 +94,7 @@ int lichen_app_interface_register_sink(
 			return -EALREADY;
 		}
 		if (!s_sinks[i].used && free_idx < 0) {
-			free_idx = i;
+			free_idx = (int)i;
 		}
 	}
 	if (free_idx < 0) {
